src/pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "psycopg2-binary>=2.9.9",
    "redis>=5.2.1",
    "gunicorn>=23.0.0",
    "python-decouple>=3.8",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "pillow>=11.0.0",
    "langgraph>=0.6.7",
    "langchain>=0.3.27",
    "google-genai>=1.36.0",
    "langchain-openai>=0.3.33",
    "langchain-google-genai>=2.1.10",
    "dotenv>=0.9.9",
    "dj-database-url>=3.0.1",
    "drf-spectacular>=0.28.0",
    "langchain-groq>=0.3.8",
    "langchain-tavily>=0.2.16",
    "celery>=5.6.2",
]
